+++ fe-prompt-mate/app/(afterlogin)/prompts/page.tsx
@@ -1,5 +1,5 @@
-import {EvaluationSectionHeader} from "@/components/EvaluationSectionHeader";
-import {PromptEvaluationCard} from "@/components/PromptEvaluationCard";
+import {EvaluationSectionHeader} from "@/components/rewrite/EvaluationSectionHeader";
+import {PromptEvaluationCard} from "@/components/rewrite/PromptEvaluationCard";
 import RewriteText from "@/components/rewrite/RewriteText";
 
 export default function PromptsPage() {
