+++ fe-prompt-mate/components/rewrite/EvaluationDetails.tsx
